requirements.txt
asgiref==3.4.1
Django==3.2.9
django-active-link==0.1.8
django-ckeditor==6.2.0
django-ckeditor-5==0.0.15
django-crispy-forms==1.13.0
django-cuser==2017.3.16
django-environ==0.8.1
django-js-asset==1.2.2
django-makemessages-xgettext==0.1.1
django-modeladmin-reorder==0.3.1
django-modeltranslation==0.17.5
django-tinymce==3.4.0
gunicorn==20.1.0
pathlib==1.0.1
Pillow==8.4.0
psycopg2-binary==2.9.3
python-decouple==3.6
python-gettext==4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
